Update testing fixtures and upgrade dependencies
This commit includes an upgrade of PHP and PHP packages in composer.json and updates to various test classes to use attributes in PHP 8.1. It also includes updates to assert methods in various test methods and changes to the structure of the code to improve readibility. The configuration for PHPUnit and PHP-CS-Fixer has also been updated reflecting these changes. Furthermore, it introduces stricter configurations, all aimed at improving quality and maintainability of the project.