Update GitHub Actions workflow to test PHP 8.1, 8.2, and 8.3
This commit updates the CI script to test against the latest PHP versions, shifting the focus from PHP 8 to PHP 8.1, 8.2, and 8.3. It includes adjustments to dependencies and testing instructions in the workflow as well. The overall goal is to maintain compatibility with recent PHP versions and ensure the reliability and quality of the project by testing it against these latest releases.